changer le type de quantite en double

+++ app/src/main/java/mg/geit/jason/Produit.kt
@@ -1,3 +1,3 @@
 package mg.geit.jason
 
-data class Produit(val id: Int?, val name: String, var prix: Double?, var quantite: Int?, var imageUrl: String, var description: String, val idCategory: Int?)
+data class Produit(val id: Int?, val name: String, var prix: Double?, var quantite: Double?, var imageUrl: String, var description: String, val idCategory: Int?)
